Just figuring things out with the viewmodel, not done yet
